+++ src/DurableTask.SqlServer/Scripts/logic.sql
@@ -1,6 +1,85 @@
 ﻿-- Copyright (c) Microsoft Corporation.
 -- Licensed under the MIT License.
 
+-- Create custom types. This must be done before creating stored procedures.
+-- IMPORTANT: If you make any changes to these types, you must also add a line in the schema upgrade script
+--            (for example, schema-1.X.0) to drop them so that those changes are properly reflected.
+IF TYPE_ID(N'__SchemaNamePlaceholder__.InstanceIDs') IS NULL
+    CREATE TYPE __SchemaNamePlaceholder__.InstanceIDs AS TABLE (
+        [InstanceID] varchar(100) NOT NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.MessageIDs') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.MessageIDs AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [SequenceNumber] bigint NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.HistoryEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.HistoryEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [SequenceNumber] bigint NULL,
+        [EventType] varchar(40) NULL,
+        [Name] varchar(300) NULL,
+        [RuntimeStatus] varchar(30) NULL,
+        [TaskID] int NULL,
+        [Timestamp] datetime2 NULL,
+        [IsPlayed] bit NULL,
+        [VisibleTime] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [ParentInstanceID] varchar(100) NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.OrchestrationEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.OrchestrationEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [SequenceNumber] bigint NULL,
+        [EventType] varchar(40) NULL,
+        [Name] varchar(300) NULL,
+        [RuntimeStatus] varchar(30) NULL,
+        [TaskID] int NULL,
+        [VisibleTime] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [ParentInstanceID] varchar(100) NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.TaskEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.TaskEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [Name] varchar(300) NULL,
+        [EventType] varchar(40) NULL,
+        [TaskID] int NULL,
+        [VisibleTime] datetime2 NULL,
+        [LockedBy] varchar(100) NULL,
+        [LockExpiration] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+
 CREATE OR ALTER FUNCTION __SchemaNamePlaceholder__.CurrentTaskHub()
     RETURNS varchar(50)
     WITH EXECUTE AS CALLER
@@ -38,12 +117,12 @@ BEGIN
     DECLARE @taskHub varchar(50) = __SchemaNamePlaceholder__.CurrentTaskHub()
     DECLARE @now datetime2 = SYSUTCDATETIME()
 
-    DECLARE @liveInstances int = 0
-    DECLARE @liveTasks int = 0
+    DECLARE @liveInstances bigint = 0
+    DECLARE @liveTasks bigint = 0
 
     SELECT
-        @liveInstances = COUNT(DISTINCT E.[InstanceID]),
-        @liveTasks = COUNT(T.[InstanceID])
+        @liveInstances = COUNT_BIG(DISTINCT E.[InstanceID]),
+        @liveTasks = COUNT_BIG(T.[InstanceID])
     FROM Instances I WITH (NOLOCK)
         LEFT OUTER JOIN NewEvents E WITH (NOLOCK) ON E.[TaskHub] = @taskHub AND E.[InstanceID] = I.[InstanceID]
         LEFT OUTER JOIN NewTasks T WITH (NOLOCK) ON T.[TaskHub] = @taskHub AND T.[InstanceID] = I.[InstanceID]
@@ -65,12 +144,12 @@ BEGIN
     DECLARE @taskHub varchar(50) = __SchemaNamePlaceholder__.CurrentTaskHub()
     DECLARE @now datetime2 = SYSUTCDATETIME()
 
-    DECLARE @liveInstances int = 0
-    DECLARE @liveTasks int = 0
+    DECLARE @liveInstances bigint = 0
+    DECLARE @liveTasks bigint = 0
 
     SELECT
-        @liveInstances = COUNT(DISTINCT E.[InstanceID]),
-        @liveTasks = COUNT(T.[InstanceID])
+        @liveInstances = COUNT_BIG(DISTINCT E.[InstanceID]),
+        @liveTasks = COUNT_BIG(T.[InstanceID])
     FROM Instances I WITH (NOLOCK)
         LEFT OUTER JOIN NewEvents E WITH (NOLOCK) ON E.[TaskHub] = @taskHub AND E.[InstanceID] = I.[InstanceID]
         LEFT OUTER JOIN NewTasks T WITH (NOLOCK) ON T.[TaskHub] = @taskHub AND T.[InstanceID] = I.[InstanceID]
@@ -102,6 +181,7 @@ AS
         I.[LastUpdatedTime],
         I.[CompletedTime],
         I.[RuntimeStatus],
+        I.[TraceContext],
         (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = __SchemaNamePlaceholder__.CurrentTaskHub() AND
             P.[InstanceID] = I.[InstanceID] AND
@@ -134,6 +214,7 @@ AS
 	    H.[Name],
 	    H.[RuntimeStatus],
         H.[VisibleTime],
+        H.[TraceContext],
 	    (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = __SchemaNamePlaceholder__.CurrentTaskHub() AND
             P.[InstanceID] = H.[InstanceID] AND
@@ -151,7 +232,8 @@ CREATE OR ALTER PROCEDURE __SchemaNamePlaceholder__.CreateInstance
     @ExecutionID varchar(50) = NULL,
     @InputText varchar(MAX) = NULL,
     @StartTime datetime2 = NULL,
-    @DedupeStatuses varchar(MAX) = 'Pending,Running'
+    @DedupeStatuses varchar(MAX) = 'Pending,Running',
+    @TraceContext varchar(800) = NULL
 AS
 BEGIN
     DECLARE @TaskHub varchar(50) = __SchemaNamePlaceholder__.CurrentTaskHub()
@@ -218,7 +300,8 @@ BEGIN
         [InstanceID],
         [ExecutionID],
         [RuntimeStatus],
-        [InputPayloadID])
+        [InputPayloadID],
+        [TraceContext])
     VALUES (
         @Name,
         @Version,
@@ -226,7 +309,8 @@ BEGIN
         @InstanceID,
         @ExecutionID,
         @RuntimeStatus,
-        @InputPayloadID
+        @InputPayloadID,
+        @TraceContext
     )
 
     INSERT INTO NewEvents (
@@ -237,6 +321,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [EventType],
+        [TraceContext],
         [PayloadID]
     ) VALUES (
         @Name,
@@ -246,6 +331,7 @@ BEGIN
         @RuntimeStatus,
         @StartTime,
         @EventType,
+        @TraceContext,
         @InputPayloadID)
 
     COMMIT TRANSACTION
@@ -282,7 +368,8 @@ BEGIN
         (CASE WHEN @GetInputsAndOutputs = 0 THEN NULL ELSE P.[Text] END) AS [PayloadText],
         [PayloadID],
         @ParentInstanceID as [ParentInstanceID],
-        @Version as [Version]
+        @Version as [Version],
+        H.[TraceContext]
     FROM History H WITH (INDEX (PK_History))
         LEFT OUTER JOIN Payloads P ON
             P.[TaskHub] = @TaskHub AND
@@ -566,7 +653,8 @@ BEGIN
         P.[PayloadID],
         DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime],
         @parentInstanceID as [ParentInstanceID],
-        @version as [Version]
+        @version as [Version],
+        N.[TraceContext]
     FROM NewEvents N
         LEFT OUTER JOIN __SchemaNamePlaceholder__.[Payloads] P ON 
             P.[TaskHub] = @TaskHub AND
@@ -605,7 +693,8 @@ BEGIN
         (CASE WHEN [EventType] IN ('TaskScheduled', 'SubOrchestrationInstanceCreated') THEN NULL ELSE P.[Text] END) AS [PayloadText],
         [PayloadID],
         @parentInstanceID as [ParentInstanceID],
-        @version as [Version]
+        @version as [Version],
+        H.[TraceContext]
     FROM History H WITH (INDEX (PK_History))
         LEFT OUTER JOIN Payloads P ON
             P.[TaskHub] = @TaskHub AND
@@ -757,14 +846,16 @@ BEGIN
         [ExecutionID],
         [Name],
         [Version],
-        [RuntimeStatus])
+        [RuntimeStatus],
+        [TraceContext])
     SELECT DISTINCT
         @TaskHub,
         E.[InstanceID],
         NEWID(),
         SUBSTRING(E.[InstanceID], 2, CHARINDEX('@', E.[InstanceID], 2) - 2),
         '',
-        'Pending'
+        'Pending',
+        E.[TraceContext]
     FROM @NewOrchestrationEvents E
     WHERE LEFT(E.[InstanceID], 1) = '@'
         AND CHARINDEX('@', E.[InstanceID], 2) > 0
@@ -772,7 +863,7 @@ BEGIN
             SELECT 1
             FROM Instances I
             WHERE [TaskHub] = @TaskHub AND I.[InstanceID] = E.[InstanceID])
-    GROUP BY E.[InstanceID]
+    GROUP BY E.[InstanceID], E.[TraceContext]
     ORDER BY E.[InstanceID] ASC
 
     -- Create sub-orchestration instances
@@ -783,7 +874,8 @@ BEGIN
         [Name],
         [Version],
         [ParentInstanceID],
-        [RuntimeStatus])
+        [RuntimeStatus],
+        [TraceContext])
     SELECT DISTINCT
         @TaskHub,
         E.[InstanceID],
@@ -791,7 +883,8 @@ BEGIN
         E.[Name],
         E.[Version],
         E.[ParentInstanceID],
-        'Pending'
+        'Pending',
+        E.[TraceContext]
     FROM @NewOrchestrationEvents E
     WHERE E.[EventType] IN ('ExecutionStarted')
         AND NOT EXISTS (
@@ -822,6 +915,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     ) 
     SELECT 
@@ -833,6 +927,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     FROM @NewOrchestrationEvents
     
@@ -861,6 +956,7 @@ BEGIN
         [Name],
         [RuntimeStatus],
         [VisibleTime],
+        [TraceContext],
         [DataPayloadID])
     SELECT
         @TaskHub,
@@ -874,6 +970,7 @@ BEGIN
         H.[Name],
         H.[RuntimeStatus],
         H.[VisibleTime],
+        H.[TraceContext],
         H.[PayloadID]
     FROM @NewHistoryEvents H
 
@@ -888,7 +985,8 @@ BEGIN
         [LockedBy],
         [LockExpiration],
         [PayloadID],
-        [Version]
+        [Version],
+        [TraceContext]
     )
     OUTPUT
         INSERTED.[SequenceNumber],
@@ -903,7 +1001,8 @@ BEGIN
         [LockedBy],
         [LockExpiration],
         [PayloadID],
-        [Version]
+        [Version],
+        [TraceContext]
     FROM @NewTaskEvents
 
     COMMIT TRANSACTION
@@ -958,20 +1057,22 @@ BEGIN
             [ExecutionID],
             [Name],
             [Version],
-            [RuntimeStatus])
+            [RuntimeStatus],
+            [TraceContext])
         SELECT DISTINCT
             @TaskHub,
             E.[InstanceID],
             NEWID(),
             SUBSTRING(E.[InstanceID], 2, CHARINDEX('@', E.[InstanceID], 2) - 2),
             '',
-            'Pending'
+            'Pending',
+            E.[TraceContext]
         FROM @NewOrchestrationEvents E
         WHERE NOT EXISTS (
             SELECT 1
             FROM Instances I
             WHERE [TaskHub] = @TaskHub AND I.[InstanceID] = E.[InstanceID])
-        GROUP BY E.[InstanceID]
+        GROUP BY E.[InstanceID], E.[TraceContext]
         ORDER BY E.[InstanceID] ASC
     END TRY
     BEGIN CATCH
@@ -998,6 +1099,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     ) 
     SELECT 
@@ -1009,6 +1111,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     FROM @NewOrchestrationEvents
 
@@ -1046,7 +1149,8 @@ BEGIN
         CASE WHEN @FetchOutput = 1 THEN (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = I.[InstanceID] AND
-            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText]
+            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText],
+        I.[TraceContext]
     FROM Instances I
     WHERE
         I.[TaskHub] = @TaskHub AND
@@ -1058,7 +1162,7 @@ GO
 
 CREATE OR ALTER PROCEDURE __SchemaNamePlaceholder__._QueryManyOrchestrations
     @PageSize smallint = 100,
-    @PageNumber smallint = 0,
+    @PageNumber int = 0,
     @FetchInput bit = 1,
     @FetchOutput bit = 1,
     @CreatedTimeFrom datetime2 = NULL,
@@ -1091,7 +1195,8 @@ BEGIN
         CASE WHEN @FetchOutput = 1 THEN (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = I.[InstanceID] AND
-            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText]
+            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText],
+        I.[TraceContext]
     FROM
         Instances I
     WHERE
@@ -1154,7 +1259,8 @@ BEGIN
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = N.[InstanceID] AND
             P.[PayloadID] = N.[PayloadID]) AS [PayloadText],
-        DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime]
+        DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime],
+        [TraceContext]
     FROM NewTasks N
     WHERE [TaskHub] = @TaskHub AND [SequenceNumber] = @SequenceNumber