+++ src/DurableTask.SqlServer/SqlUtils.cs
@@ -10,10 +10,12 @@ namespace DurableTask.SqlServer
     using System.Data.SqlTypes;
     using System.Diagnostics;
     using System.Linq;
+    using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
     using DurableTask.Core;
     using DurableTask.Core.History;
+    using DurableTask.Core.Tracing;
     using Microsoft.Data.SqlClient;
     using Microsoft.Data.SqlClient.Server;
     using SemVersion;
@@ -21,6 +23,7 @@ namespace DurableTask.SqlServer
     static class SqlUtils
     {
         static readonly Random random = new Random();
+        static readonly char[] TraceContextSeparators = new char[] { '\n' };
 
         public static string? GetStringOrNull(this DbDataReader reader, int columnIndex)
         {
@@ -61,6 +64,7 @@ namespace DurableTask.SqlServer
                     historyEvent = new EventRaisedEvent(eventId, GetPayloadText(reader))
                     {
                         Name = GetName(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     break;
                 case EventType.EventSent:
@@ -78,10 +82,18 @@ namespace DurableTask.SqlServer
                         orchestrationStatus: OrchestrationStatus.Completed);
                     break;
                 case EventType.ExecutionFailed:
+                    string? executionFailedResult = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? executionFailedDetails))
+                    {
+                        // Fall back to the old behavior
+                        executionFailedResult = GetPayloadText(reader);
+                    }
+
                     historyEvent = new ExecutionCompletedEvent(
                         eventId,
-                        result: GetPayloadText(reader),
-                        orchestrationStatus: OrchestrationStatus.Failed);
+                        result: executionFailedResult,
+                        orchestrationStatus: OrchestrationStatus.Failed,
+                        failureDetails: executionFailedDetails);
                     break;
                 case EventType.ExecutionStarted:
                     historyEvent = new ExecutionStartedEvent(eventId, GetPayloadText(reader))
@@ -94,6 +106,7 @@ namespace DurableTask.SqlServer
                         },
                         Tags = null, // TODO
                         Version = GetVersion(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     string? parentInstanceId = GetParentInstanceId(reader);
                     if (parentInstanceId != null)
@@ -133,11 +146,21 @@ namespace DurableTask.SqlServer
                     };
                     break;
                 case EventType.SubOrchestrationInstanceFailed:
+                    string? subOrchFailedReason = null;
+                    string? subOrchFailedDetails = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? subOrchFailureDetails))
+                    {
+                        // Fall back to the old behavior
+                        subOrchFailedReason = GetReason(reader);
+                        subOrchFailedDetails = GetPayloadText(reader);
+                    }
+
                     historyEvent = new SubOrchestrationInstanceFailedEvent(
                         eventId,
                         taskScheduledId: GetTaskId(reader),
-                        reason: GetReason(reader),
-                        details: GetPayloadText(reader));
+                        reason: subOrchFailedReason,
+                        details: subOrchFailedDetails,
+                        failureDetails: subOrchFailureDetails);
                     break;
                 case EventType.TaskCompleted:
                     historyEvent = new TaskCompletedEvent(
@@ -146,11 +169,21 @@ namespace DurableTask.SqlServer
                         result: GetPayloadText(reader));
                     break;
                 case EventType.TaskFailed:
+                    string? taskFailedReason = null;
+                    string? taskFailedDetails = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? taskFailureDetails))
+                    {
+                        // Fall back to the old behavior
+                        taskFailedReason = GetReason(reader);
+                        taskFailedDetails = GetPayloadText(reader);
+                    }
+
                     historyEvent = new TaskFailedEvent(
                         eventId,
                         taskScheduledId: GetTaskId(reader),
-                        reason: GetReason(reader),
-                        details: GetPayloadText(reader));
+                        reason: taskFailedReason,
+                        details: taskFailedDetails,
+                        failureDetails: taskFailureDetails);
                     break;
                 case EventType.TaskScheduled:
                     historyEvent = new TaskScheduledEvent(eventId)
@@ -158,6 +191,7 @@ namespace DurableTask.SqlServer
                         Input = GetPayloadText(reader),
                         Name = GetName(reader),
                         Version = GetVersion(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     break;
                 case EventType.TimerCreated:
@@ -182,6 +216,20 @@ namespace DurableTask.SqlServer
             return historyEvent;
         }
 
+        static bool TryGetFailureDetails(DbDataReader reader, out FailureDetails? details)
+        {
+            // Failure details are expected to be in JSON format. In previous versions, it was just an error message
+            // that isn't expected to be JSON.
+            string? text = GetPayloadText(reader);
+            if (string.IsNullOrEmpty(text) || text![0] != '{')
+            {
+                details = null;
+                return false;
+            }
+
+            return DTUtils.TryDeserializeFailureDetails(text, out details);
+        }
+
         public static OrchestrationState GetOrchestrationState(this DbDataReader reader)
         {
             ParentInstance? parentInstance = null;
@@ -196,7 +244,8 @@ namespace DurableTask.SqlServer
                     }
                 };
             }
-            return new OrchestrationState
+
+            var state = new OrchestrationState
             {
                 CompletedTime = reader.GetUtcDateTimeOrNull(reader.GetOrdinal("CompletedTime")) ?? default,
                 CreatedTime = reader.GetUtcDateTimeOrNull(reader.GetOrdinal("CreatedTime")) ?? default,
@@ -210,10 +259,27 @@ namespace DurableTask.SqlServer
                     ExecutionId = GetExecutionId(reader),
                 },
                 OrchestrationStatus = GetRuntimeStatus(reader),
-                Output = GetStringOrNull(reader, reader.GetOrdinal("OutputText")),
                 Status = GetStringOrNull(reader, reader.GetOrdinal("CustomStatusText")),
                 ParentInstance = parentInstance
             };
+
+            // The OutputText column is overloaded to contain either orchestration output or failure details
+            // if the task hub worker is configured to use legacy error propagation.
+            string? rawOutput = GetStringOrNull(reader, reader.GetOrdinal("OutputText"));
+            if (rawOutput != null)
+            {
+                if (state.OrchestrationStatus == OrchestrationStatus.Failed &&
+                    DTUtils.TryDeserializeFailureDetails(rawOutput, out FailureDetails? failureDetails))
+                {
+                    state.FailureDetails = failureDetails;
+                }
+                else
+                {
+                    state.Output = rawOutput;
+                }
+            }
+
+            return state;
         }
 
         internal static DateTime? GetVisibleTime(HistoryEvent historyEvent)
@@ -364,6 +430,53 @@ namespace DurableTask.SqlServer
         {
             // The SQL client always assumes DateTimeKind.Unspecified. We need to modify the result so that it knows it is UTC.
             return DateTime.SpecifyKind(reader.GetDateTime(ordinal), DateTimeKind.Utc);
+        }
+
+        internal static SqlString GetTraceContext(HistoryEvent e)
+        {
+            if (e is not ISupportsDurableTraceContext eventWithTraceContext ||
+                eventWithTraceContext.ParentTraceContext == null)
+            {
+                return SqlString.Null;
+            }
+
+            DistributedTraceContext traceContext = eventWithTraceContext.ParentTraceContext;
+
+            // We prefer a simple format instead of JSON because external callers may interact with this
+            // data and we don't want to expose them to some internal JSON serialization format.
+            var sb = new StringBuilder(traceContext.TraceParent, capacity: 800);
+            if (!string.IsNullOrEmpty(traceContext.TraceState))
+            {
+                sb.Append('\n').Append(traceContext.TraceState);
+            }
+
+            return sb.ToString();
+        }
+
+        static DistributedTraceContext? GetTraceContext(DbDataReader reader)
+        {
+            int ordinal = reader.GetOrdinal("TraceContext");
+            if (reader.IsDBNull(ordinal))
+            {
+                return null;
+            }
+
+            string text = reader.GetString(ordinal);
+            if (string.IsNullOrEmpty(text))
+            {
+                return null;
+            }
+
+            string[] parts = text.Split(TraceContextSeparators, count: 2, StringSplitOptions.RemoveEmptyEntries);
+            var traceContext = new DistributedTraceContext(traceParent: parts[0]);
+
+            if (parts.Length > 1)
+            {
+                traceContext.TraceState = parts[1];
+            }
+
+            traceContext.ActivityStartTime = GetTimestamp(reader);
+            return traceContext;
         }
 
         public static SqlParameter AddInstanceIDsParameter(