+++ test/setup.ps1
@@ -16,7 +16,7 @@ docker pull mcr.microsoft.com/mssql/server:$tag
 
 # Start the SQL Server docker container with the specified edition
 Write-Host "Starting SQL Server $tag $sqlpid docker container on port $port" -ForegroundColor DarkYellow
-docker run $additinalRunFlags --name mssql-server -e 'ACCEPT_EULA=Y' -e "SA_PASSWORD=$pw" -e "MSSQL_PID=$sqlpid" -p ${port}:1433 -d mcr.microsoft.com/mssql/server:$tag
+docker run $additinalRunFlags --name mssql-server -e 'ACCEPT_EULA=Y' -e "MSSQL_SA_PASSWORD=$pw" -e "MSSQL_PID=$sqlpid" -p ${port}:1433 -d mcr.microsoft.com/mssql/server:$tag
 
 # The container needs a bit more time before it can start accepting commands
 Write-Host "Sleeping for 30 seconds to let the container finish initializing..." -ForegroundColor Yellow